// src/api/client/api.ts
export interface ClientResponse {
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  phone?: string;
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  email?: string;
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  website?: string;
  /**
   *
   * @type {string}
   * @memberof ClientResponse
   */
  observation?: string;
}
/**
 *
 * @export
 * @interface CreateClientRequest
 */
export interface CreateClientRequest {
  /**
   *
   * @type {string}
   * @memberof CreateClientRequest
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof CreateClientRequest
   */
  email: string;
  /**
   *
   * @type {string}
   * @memberof CreateClientRequest
   */
  phone?: string;
  /**
   *
   * @type {string}
   * @memberof CreateClientRequest
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof CreateClientRequest
   */
  website?: string;
  /**
   *
   * @type {string}
   * @memberof CreateClientRequest
   */
  observation?: string;
}
/**
 *
 * @export
 * @interface ErrorResponse
 */
export interface ErrorResponse {
  /**
   *
   * @type {string}
   * @memberof ErrorResponse
   */
  error?: string;
  /**
   *
   * @type {number}
   * @memberof ErrorResponse
   */
  code?: number;
}
/**
 *
 * @export
 * @interface GetAllClientsPaginatedResponse
 */
export interface GetAllClientsPaginatedResponse {
  /**
   *
   * @type {Pagination}
   * @memberof GetAllClientsPaginatedResponse
   */
  pagination?: Pagination;
  /**
   *
   * @type {Array<ClientResponse>}
   * @memberof GetAllClientsPaginatedResponse
   */
  data?: Array<ClientResponse>;
}
/**
 *
 * @export
 * @interface Pagination
 */
export interface Pagination {
  /**
   *
   * @type {number}
   * @memberof Pagination
   */
  totalItems?: number;
  /**
   *
   * @type {number}
   * @memberof Pagination
   */
  pageSize?: number;
  /**
   *
   * @type {number}
   * @memberof Pagination
   */
  currentPage?: number;
  /**
   *
   * @type {number}
   * @memberof Pagination
   */
  totalPages?: number;
}
/**
 *
 * @export
 * @interface UpdateClientRequest
 */
export interface UpdateClientRequest {
  /**
   *
   * @type {string}
   * @memberof UpdateClientRequest
   */
  name?: string;
  /**
   *
   * @type {string}
   * @memberof UpdateClientRequest
   */
  email?: string;
  /**
   *
   * @type {string}
   * @memberof UpdateClientRequest
   */
  phone?: string;
  /**
   *
   * @type {string}
   * @memberof UpdateClientRequest
   */
  address?: string;
  /**
   *
   * @type {string}
   * @memberof UpdateClientRequest
   */
  website?: string;
  /**
   *
   * @type {string}
   * @memberof UpdateClientRequest
   */
  observation?: string;
}
// ...
export interface ClientsApiCreateClientRequest {
  /**
   *
   * @type {CreateClientRequest}
   * @memberof ClientsApiCreateClient
   */
  readonly createClientRequest: CreateClientRequest;
}

/**
 * Request parameters for deleteClient operation in ClientsApi.
 * @export
 * @interface ClientsApiDeleteClientRequest
 */
export interface ClientsApiDeleteClientRequest {
  /**
   * Unique identifier of the client to delete
   * @type {string}
   * @memberof ClientsApiDeleteClient
   */
  readonly id: string;
}

/**
 * Request parameters for getClients operation in ClientsApi.
 * @export
 * @interface ClientsApiGetClientsRequest
 */
export interface ClientsApiGetClientsRequest {
  /**
   * The starting index of the invoices.
   * @type {number}
   * @memberof ClientsApiGetClients
   */
  readonly from?: number;

  /**
   * The number of invoices to return.
   * @type {number}
   * @memberof ClientsApiGetClients
   */
  readonly size?: number;

  /**
   * Filter criteria for invoices.
   * @type {string}
   * @memberof ClientsApiGetClients
   */
  readonly filter?: string;

  /**
   * Field to order clients by.
   * @type {'createdAt'}
   * @memberof ClientsApiGetClients
   */
  readonly orderBy?: GetClientsOrderByEnum;

  /**
   * Sorting direction.
   * @type {'ASC' | 'DESC'}
   * @memberof ClientsApiGetClients
   */
  readonly direction?: GetClientsDirectionEnum;

  /**
   * Filter clients by their status.
   * @type {'ACTIVE' | 'INACTIVE'}
   * @memberof ClientsApiGetClients
   */
  readonly status?: GetClientsStatusEnum;
}

/**
 * Request parameters for updateClient operation in ClientsApi.
 * @export
 * @interface ClientsApiUpdateClientRequest
 */
export interface ClientsApiUpdateClientRequest {
  /**
   * Unique identifier of the client to update
   * @type {string}
   * @memberof ClientsApiUpdateClient
   */
  readonly id: string;

  /**
   *
   * @type {UpdateClientRequest}
   * @memberof ClientsApiUpdateClient
   */
  readonly updateClientRequest: UpdateClientRequest;
}

/**
 * ClientsApi - object-oriented interface
 * @export
 * @class ClientsApi
 * @extends {BaseAPI}
 */
export class ClientsApi extends BaseAPI {
  /**
   *
   * @summary Create a new client
   * @param {ClientsApiCreateClientRequest} requestParameters Request parameters.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ClientsApi
   */
  public createClient(requestParameters: ClientsApiCreateClientRequest, options?: RawAxiosRequestConfig) {
    return ClientsApiFp(this.configuration)
      .createClient(requestParameters.createClientRequest, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Delete a client by ID
   * @param {ClientsApiDeleteClientRequest} requestParameters Request parameters.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ClientsApi
   */
  public deleteClient(requestParameters: ClientsApiDeleteClientRequest, options?: RawAxiosRequestConfig) {
    return ClientsApiFp(this.configuration)
      .deleteClient(requestParameters.id, options)
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Retrieve a list of clients
   * @param {ClientsApiGetClientsRequest} requestParameters Request parameters.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ClientsApi
   */
  public getClients(requestParameters: ClientsApiGetClientsRequest = {}, options?: RawAxiosRequestConfig) {
    return ClientsApiFp(this.configuration)
      .getClients(
        requestParameters.from,
        requestParameters.size,
        requestParameters.filter,
        requestParameters.orderBy,
        requestParameters.direction,
        requestParameters.status,
        options
      )
      .then((request) => request(this.axios, this.basePath));
  }

  /**
   *
   * @summary Update an existing client
   * @param {ClientsApiUpdateClientRequest} requestParameters Request parameters.
   * @param {*} [options] Override http request option.
   * @throws {RequiredError}
   * @memberof ClientsApi
   */
  public updateClient(requestParameters: ClientsApiUpdateClientRequest, options?: RawAxiosRequestConfig) {
    return ClientsApiFp(this.configuration)
      .updateClient(requestParameters.id, requestParameters.updateClientRequest, options)
      .then((request) => request(this.axios, this.basePath));
  }
}

/**
 * @export
 */
export const GetClientsOrderByEnum = {
  CreatedAt: 'createdAt',
} as const;
export type GetClientsOrderByEnum = (typeof GetClientsOrderByEnum)[keyof typeof GetClientsOrderByEnum];
/**
 * @export
 */
export const GetClientsDirectionEnum = {
  Asc: 'ASC',
  Desc: 'DESC',
} as const;
export type GetClientsDirectionEnum = (typeof GetClientsDirectionEnum)[keyof typeof GetClientsDirectionEnum];
/**
 * @export
 */
export const GetClientsStatusEnum = {
  Active: 'ACTIVE',
  Inactive: 'INACTIVE',
} as const;
export type GetClientsStatusEnum = (typeof GetClientsStatusEnum)[keyof typeof GetClientsStatusEnum];